+++ cli/cli/helpers/engine_manager/engine_manager.go
@@ -43,50 +43,7 @@ Returns:
 func (manager *EngineManager) GetEngineStatus(
 	ctx context.Context,
 ) (EngineStatus, *nat.PortBinding, string, error) {
-	runningEngineContainers, err := manager.dockerManager.GetContainersByLabels(ctx, engine_labels_schema.EngineContainerLabels, shouldGetStoppedContainersWhenCheckingForExistingEngines)
-	if err != nil {
-		return "", nil, "", stacktrace.Propagate(err, "An error occurred getting Kurtosis engine containers")
-	}
-
-	numRunningEngineContainers := len(runningEngineContainers)
-	if numRunningEngineContainers > 1 {
-		return "", nil, "", stacktrace.NewError("Cannot report engine status because we found %v running Kurtosis engine containers; this is very strange as there should never be more than one", numRunningEngineContainers)
-	}
-	if numRunningEngineContainers == 0 {
-		return EngineStatus_Stopped, nil, "", nil
-	}
-	engineContainer := runningEngineContainers[0]
-
-	enginePortObj, err := nat.NewPort(
-		kurtosis_engine_rpc_api_consts.ListenProtocol,
-		fmt.Sprintf("%v", kurtosis_engine_rpc_api_consts.ListenPort),
-	)
-	if err != nil {
-		return "", nil, "", stacktrace.Propagate(
-			err,
-			"An error occurred creating an engine port object from port num '%v' and protocol '%v'",
-			kurtosis_engine_rpc_api_consts.ListenPort,
-			kurtosis_engine_rpc_api_consts.ListenProtocol,
-		)
-	}
-
-	hostMachineEnginePortBinding, found := engineContainer.GetHostPortBindings()[enginePortObj]
-	if !found {
-		return "", nil, "", stacktrace.NewError("Found a Kurtosis engine server container, but it didn't have a host machine port binding - this is likely a Kurtosis bug")
-	}
-
-	engineClient, clientCloseFunc, err := getEngineClientFromHostPortBinding(hostMachineEnginePortBinding)
-	if err != nil {
-		return EngineStatus_ContainerRunningButServerNotResponding, hostMachineEnginePortBinding, "", nil
-	}
-	defer clientCloseFunc()
-
-	engineInfo, err := getEngineInfoWithTimeout(ctx, engineClient)
-	if err != nil {
-		return EngineStatus_ContainerRunningButServerNotResponding, hostMachineEnginePortBinding, "", nil
-	}
-
-	return EngineStatus_Running, hostMachineEnginePortBinding, engineInfo.EngineApiVersion, nil
+	return getEngineStatus(ctx, manager.dockerManager)
 }
 
 // Starts an engine if one doesn't exist already, and returns a client to it
@@ -191,3 +148,53 @@ func getEngineInfoWithTimeout(ctx context.Context, client kurtosis_engine_rpc_ap
 	return engineInfo, nil
 }
 
+func getEngineStatus(ctx context.Context, dockerManager *docker_manager.DockerManager) (EngineStatus, *nat.PortBinding, string, error) {
+	runningEngineContainers, err := dockerManager.GetContainersByLabels(ctx, engine_labels_schema.EngineContainerLabels, shouldGetStoppedContainersWhenCheckingForExistingEngines)
+	if err != nil {
+		return "", nil, "", stacktrace.Propagate(err, "An error occurred getting Kurtosis engine containers")
+	}
+
+	numRunningEngineContainers := len(runningEngineContainers)
+	if numRunningEngineContainers > 1 {
+		return "", nil, "", stacktrace.NewError("Cannot report engine status because we found %v running Kurtosis engine containers; this is very strange as there should never be more than one", numRunningEngineContainers)
+	}
+	if numRunningEngineContainers == 0 {
+		return EngineStatus_Stopped, nil, "", nil
+	}
+	engineContainer := runningEngineContainers[0]
+
+	enginePortObj, err := nat.NewPort(
+		kurtosis_engine_rpc_api_consts.ListenProtocol,
+		fmt.Sprintf("%v", kurtosis_engine_rpc_api_consts.ListenPort),
+	)
+	if err != nil {
+		return "", nil, "", stacktrace.Propagate(
+			err,
+			"An error occurred creating an engine port object from port num '%v' and protocol '%v'",
+			kurtosis_engine_rpc_api_consts.ListenPort,
+			kurtosis_engine_rpc_api_consts.ListenProtocol,
+		)
+	}
+
+	hostMachineEnginePortBinding, found := engineContainer.GetHostPortBindings()[enginePortObj]
+	if !found {
+		return "", nil, "", stacktrace.NewError("Found a Kurtosis engine server container, but it didn't have a host machine port binding - this is likely a Kurtosis bug")
+	}
+
+	engineClient, clientCloseFunc, err := getEngineClientFromHostPortBinding(hostMachineEnginePortBinding)
+	if err != nil {
+		return EngineStatus_ContainerRunningButServerNotResponding, hostMachineEnginePortBinding, "", nil
+	}
+	defer func() {
+		if err := clientCloseFunc(); err != nil {
+			logrus.Warnf("We tried to close the engine client, but doing so threw an error:\n%v", err)
+		}
+	}()
+
+	engineInfo, err := getEngineInfoWithTimeout(ctx, engineClient)
+	if err != nil {
+		return EngineStatus_ContainerRunningButServerNotResponding, hostMachineEnginePortBinding, "", nil
+	}
+
+	return EngineStatus_Running, hostMachineEnginePortBinding, engineInfo.EngineApiVersion, nil
+}